refactor: separate PermissionList
