+++ src/components/AccountDrawer.tsx
@@ -27,11 +27,10 @@ import {
 import { AddCircleOutline, RemoveCircleOutline } from "@material-ui/icons";
 import { DarkMode, LightMode, Reload } from "components/MaterialSvgIcons";
 import AccountIcon from "components/AccountIcon";
-import PermissionIcon from "components/PermissionIcon";
+import PermissionList from "components/PermissionList";
 import { useAuthDispatch, useAuthState } from "libs/auth/useAuth";
 import { useSetThemeMode } from "libs/themeMode";
 import { Alert } from "@material-ui/lab";
-import clsx from "clsx";
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -50,9 +49,6 @@ const useStyles = makeStyles((theme: Theme) =>
     },
     currentUserPerm: {
       marginTop: theme.spacing(0.5),
-    },
-    disabledIcon: {
-      opacity: theme.palette.action.disabledOpacity,
     },
     menuIcon: {
       marginBottom: theme.spacing(1),
@@ -78,13 +74,6 @@ const useStyles = makeStyles((theme: Theme) =>
     },
     snackBar: {
       bottom: "calc(64px + env(safe-area-inset-bottom))",
-    },
-    permissionsList: {
-      display: "flex",
-      height: "min-content",
-      "& > * + *": {
-        marginLeft: theme.spacing(0.5),
-      },
     },
   })
 );
@@ -126,7 +115,7 @@ const AccountDrawer: React.VFC<Props> = ({
         <Paper className={classes.currentUser} square={true}>
           <div className={classes.currentUserIconWrapper}>
             <AccountIcon account={currentUser} className={classes.menuIcon} />
-            <PermissionsList />
+            <PermissionList account={currentUser} />
           </div>
           <Typography variant="h6">{currentUser.name}</Typography>
           <Typography variant="body2">@{currentUser.id}</Typography>
@@ -290,27 +279,6 @@ const AccountDrawer: React.VFC<Props> = ({
         </Alert>
       </Snackbar>
     </Drawer>
-  );
-};
-
-export const PermissionsList: React.VFC<{ className?: string }> = ({
-  className,
-}) => {
-  const classes = useStyles();
-  const { currentUser } = useAuthState();
-
-  if (!currentUser) return null;
-  return (
-    <div className={clsx(classes.permissionsList, className)}>
-      {Object.entries(currentUser.permissions).map(([name, val]) => (
-        <PermissionIcon
-          permName={name}
-          key={name}
-          fontSize="small"
-          className={clsx({ [classes.disabledIcon]: !val })}
-        />
-      ))}
-    </div>
   );
 };
