+++ src/pages/Account.tsx
@@ -20,7 +20,7 @@ import {
 } from "@material-ui/core";
 import { AddCircleOutline, RemoveCircleOutline } from "@material-ui/icons";
 import AccountIcon from "components/AccountIcon";
-import { PermissionsList } from "components/AccountDrawer";
+import PermissionList from "components/PermissionList";
 import { useTitleSet } from "libs/title";
 import routes from "libs/routes";
 import { Link } from "react-router-dom";
@@ -88,7 +88,10 @@ const Account: React.VFC = () => {
         />
         <Typography variant="h6">{currentUser?.name ?? ""}</Typography>
         <Typography variant="body2">@{currentUser?.id ?? ""}</Typography>
-        <PermissionsList className={classes.currentUserPerm} />
+        <PermissionList
+          account={currentUser}
+          className={classes.currentUserPerm}
+        />
       </Grid>
 
       {/* ==== account list ==== */}
